GCW-3438 Move useAuth hook to its own folder
