+++ src/components/PrivateRoute.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Redirect, Route } from "react-router";
-import useAuth from "hooks/useAuth";
+import useAuth from "hooks/useAuth/useAuth";
 
 interface RestProps {
   exact: boolean;
